replaced deprecated ebiten text with textv2

+++ v2/simulation/test/simulation_test.go
@@ -164,7 +164,7 @@ func TestJuvenilePhaseBlocksReproduction(t *testing.T) {
 
 	for _, c := range sim.Population.Creatures {
 		c.Genome.JuvenilePeriod = 255
-		c.Energy = c.Mass * p.EnergyPerMassUnit
+		c.Energy = float32(c.Mass) * p.EnergyPerMassUnit
 		c.Age = 0 // reset so the full juvenile phase must elapse before reproduction
 	}
 
@@ -197,7 +197,7 @@ func TestAdultCreaturesCanReproduce(t *testing.T) {
 		c.Genome.JuvenilePeriod = 255
 		c.Genome.ReproductionType = 0   // asexual
 		c.Genome.MassSplitRatio = 128   // ~25% split
-		c.Energy = c.Mass * p.EnergyPerMassUnit
+		c.Energy = float32(c.Mass) * p.EnergyPerMassUnit
 		// Wire ENERGY sensor directly to REPRODUCE action so it fires unconditionally.
 		c.Genome.Brain = append(c.Genome.Brain, simulation.Gene{
 			SourceType: simulation.SENSOR,
